requirements.txt
pip==22.3.1
wheel==0.38.4
setuptools==65.6.3
numpy~=1.24.1
Pillow~=9.4.0
matplotlib~=3.6.2